LC3 File consisting of encryptor.
1.  We first initialize all variables in the registers that will be used
in various routines, including the values of inputs requiring basic
arithmetic operations.

2. Once the program starts, a message prints out to the screen asking
the user to choose an option (E for encrypt, D for decrypt, or X for
exit) and handles malicious input.

3. If “E” is chosen, the program asks the user to enter a string of 16
characters, and verifies that the string is of the correct length.

4. The program will now ask for the user to enter in a key used for
encryption. Once again, the program handles malicious input, including
whether the key fits the correct bounds.

5. The least significant bit is toggled, or flipped.

6. The key is then added to each character, which will be calculated
using mod 128 (the number of ASCII values). This ensures that the
boundaries of the ASCII table have not been surpassed.

7. Having completed this, we return to the input menu, and a new process
begins.

8. If “D” is chosen, we first check if any encryption has taken place to
begin with. If not, then the program prints out a corresponding message
that an error has taken place, and we return to the menu.

9. If the conditions of (8) have been passed, the user will be asked to
enter an encryption key.

10. Using inverses and modulo, the program convert substitutes the
entered key of decryption for one that will allow the full process to
take place.

11. The modulo processes for encryption (5-6) are repeated in decryption
as well.

12. The resulting decrypted message is then printed out, and we return
to the menu yet again.

13. If “X” is entered, the program stops, or as we say in assembly
“halts”.